Consistent exporting & naming of components files
