+++ webclient/src/components/BlogList/index.js
@@ -1,0 +1,2 @@
+import BlogList from './BlogList';
+export default BlogList;
