+++ webclient/src/components/BlogListItem/index.js
@@ -1,0 +1,2 @@
+import BlogListItem from './BlogListItem';
+export default BlogListItem;
