+++ webclient/src/components/BlogPost/index.js
@@ -1,0 +1,2 @@
+import BlogPost from './BlogPost';
+export default BlogPost;
