+++ webclient/src/components/Card/index.js
@@ -1,0 +1,2 @@
+import Card from './Card';
+export default Card;
