+++ webclient/src/components/CodeList/index.js
@@ -1,0 +1,2 @@
+import CodeList from './CodeList';
+export default CodeList;
