+++ webclient/src/components/HeaderLayout/HeaderLayout.jsx
@@ -9,7 +9,7 @@ import { toggleOptionsMenu } from '../../redux/modules/ui/optionsMenu';
 
 import Button from 'react-md/lib/Buttons/Button';
 
-class Header extends React.Component {
+class HeaderLayout extends React.Component {
   render() {
     return (
       <div className="header-layout">
@@ -30,7 +30,7 @@ class Header extends React.Component {
   }
 }
 
-Header.propTypes = {
+HeaderLayout.propTypes = {
   title: PropTypes.string.isRequired,
   handleMenuButtonClick: PropTypes.func.isRequired,
   handleOptionsButtonClick: PropTypes.func.isRequired
@@ -46,4 +46,4 @@ const mapDispatchToProps = {
   handleOptionsButtonClick: toggleOptionsMenu
 }
 
-export default connect(mapStateToProps, mapDispatchToProps)(Header)
+export default connect(mapStateToProps, mapDispatchToProps)(HeaderLayout)
