+++ webclient/src/components/HeaderLayout/index.js
@@ -1,0 +1,2 @@
+import HeaderLayout from './HeaderLayout';
+export default HeaderLayout;
