+++ webclient/src/components/Icon/Icon.js
@@ -1,0 +1,8 @@
+import React from 'react';
+
+// Functional Component
+const Icon = (props) => {
+  return  <img src={props.img} alt={props.alt} title={props.alt} className={props.className}></img>
+}
+
+export default Icon;
