+++ webclient/src/components/Icon/index.js
@@ -1,8 +1,2 @@
-import React from 'react';
-
-// Functional Component
-const Icon = (props) => {
-  return  <img src={props.img} alt={props.alt} title={props.alt} className={props.className}></img>
-}
-
+import Icon from './Icon';
 export default Icon;
