+++ webclient/src/components/Link/index.js
@@ -1,0 +1,2 @@
+import Link from './Link';
+export default Link;
