+++ webclient/src/components/OptionsMenu/index.js
@@ -1,0 +1,2 @@
+import OptionsMenu from './OptionsMenu';
+export default OptionsMenu;
