+++ webclient/src/components/OptionsMenuItem/index.js
@@ -1,0 +1,2 @@
+import OptionsMenuItem from './OptionsMenuItem';
+export default OptionsMenuItem;
