+++ webclient/src/components/PersonalLinks/index.js
@@ -1,0 +1,2 @@
+import PersonalLinks from './PersonalLinks';
+export default PersonalLinks;
