+++ webclient/src/components/ScreenMediaQuery/index.js
@@ -1,0 +1,2 @@
+import ScreenMediaQuery from './ScreenMediaQuery';
+export default ScreenMediaQuery;
