+++ webclient/src/components/SideNavLayout/index.js
@@ -1,0 +1,2 @@
+import SideNavLayout from './SideNavLayout';
+export default SideNavLayout;
