+++ webclient/src/components/SideNavLink/index.js
@@ -1,0 +1,2 @@
+import SideNavLink from './SideNavLink';
+export default SideNavLink;
