+++ webclient/src/components/Spinner/index.js
@@ -1,0 +1,2 @@
+import Spinner from './Spinner';
+export default Spinner;
