+++ webclient/src/components/Toast/index.js
@@ -1,0 +1,2 @@
+import Toast from './Toast';
+export default Toast;
